[fix] warning to compare unsigned value and signed value
符号の有無による警告を受けたので符号なしに統一した

+++ ColdSoupGenerator/ColdSoupGenerator/AXFCodeSegment.cpp
@@ -13,7 +13,7 @@ AXFCodeSegment::~AXFCodeSegment()
 {
 }
 
-void AXFCodeSegment::Load(std::ifstream &axfile, int index, int size)
+void AXFCodeSegment::Load(std::ifstream &axfile, size_t index, size_t size)
 {
 	axfile.seekg(index);
 	{
